Update layout
Added an extra layout for the posts which includes a back to cameron's log button, serving as title also
Changed all posts over to this layout.
Also added a UTF-8 content type header thing to prevent the apostrophe problem.
And added a little bit of padding in the css file.